Add Copyright statement to README
Via #55. To apply the Apache License, the boilerplate notice should be
specified, and the copyright owner clarified.

The Community License Agreement that WebDX CG participants signed when they
joined the group covers Specifications and does not have anything to say about
projects such as this one. Making the copyright owner "baseline-status
contributors" as a result.

The notice could be added to each and every source file. This update keeps
things simple and adds it once to the README file instead.

The copyright year is not specified but I don't think that matters much these
days, and this avoids having to get back to it to update the years over time.